Orthogonal Regression

orthogonal_regression

ErrorsInVariables.OrthogonalRegression.orthogonal_regression — Function

orthogonal_regression(
    X::Matrix,
    y::Vector;
    xhasintercept::Bool=true,
    maxiterations::Int=10000,
    initialbetas::Union{Nothing, Vector} = nothing)::SimpleEiveResult

Description:

The function performs orthogonal regression. The method searches for a set of parameters that minimize the sum of squares of orthogonal residuals.

Arguments:

X::Matrix: Independent variables. The first column should be 1 for the intercept.

y::Vector: Dependent variable.

xhasintercept::Bool: If true, the first column of X is considered as the intercept.

maxiterations::Int: Maximum number of iterations.

initialbetas::Union{Nothing, Vector}: Initial values for the parameters. If nothing, random values are used.

Returns:

::EiveResult: A struct containing the estimated parameters and convergence status.

Examples

X = Float64[1 1.0; 1 0.6; 1 1.2; 1 1.4; 1 0.2]
y = Float64[0.5, 0.3, 0.7, 1.0, 0.2]

result = orthogonal_regression(X, y)

References

https://davegiles.blogspot.com/2014/11/orthogonal-regression-first-steps.html