Simulation Extrapolation

simex

ErrorsInVariables.SimulationExtrapolation.simex — Function

simex(
X::Matrix,
y::Vector,
λ::Vector,
errorvarindex::Int,
errvariance::Float64;
numsims::Int = 1000)::SimpleEiveResult

Description

This function implements the Simulation Extrapolation (SIMEX) method. The function fits a linear model to the data and then extrapolates the results to a new value of the parameter λ. The function returns the estimated parameters and a boolean indicating if the optimization converged.

Arguments

X::Matrix: A matrix of the independent variables.

y::Vector: A vector of the dependent variable.

λ::Vector: A vector of the values of the parameter λ.

errorvarindex::Int: The index of the error variable in the X matrix.

errvariance::Float64: The variance of the error variable.

numsims::Int = 1000: The number of simulations to run.

Returns

SimpleEiveResult: An object that contains the estimated parameters of the model.

References

Cook, John R., and Leonard A. Stefanski. "Simulation-extrapolation estimation in parametric

measurement error models." Journal of the American Statistical association 89.428 (1994): 1314-1328.

simexsingleiteration

ErrorsInVariables.SimulationExtrapolation.simex_single_iteration — Function

simex_single_iteration(
    X::Matrix, 
    y::Vector, 
    λ::Float64, 
    errorvarindex::Int, 
    errvariance::Float64; 
    numsims::Int = 1000)::Vector

Description

This function performs a single iteration of the Simulation Extrapolation (SIMEX) method.

Arguments

X::Matrix: A matrix of the independent variables.

y::Vector: A vector of the dependent variable.

λ::Float64: The value of the parameter λ.

errorvarindex::Int: The index of the error variable in the X matrix.

errvariance::Float64: The variance of the error variable.

numsims::Int = 1000: The number of simulations to run.

Returns

Vector: A vector of the estimated parameters of the model.

simexmultipleiterations

ErrorsInVariables.SimulationExtrapolation.simex_multiple_iterations — Function

simex_multiple_iterations(
    X::Matrix, 
    y::Vector, 
    λ::Vector, 
    errorvarindex::Int, 
    errvariance::Float64; 
    numsims::Int = 1000
)::Matrix

Description

This function performs multiple iterations of the Simulation Extrapolation (SIMEX) method.

Arguments

X::Matrix: A matrix of the independent variables.

y::Vector: A vector of the dependent variable.

λ::Vector: A vector of the values of the parameter λ.

errorvarindex::Int: The index of the error variable in the X matrix.

errvariance::Float64: The variance of the error variable.

numsims::Int = 1000: The number of simulations to run.

Returns

Matrix: A matrix of the estimated parameters of the model.

extrapolate

ErrorsInVariables.SimulationExtrapolation.extrapolate — Function

extrapolate(λ::Vector, betas::Matrix, errorvarindex::Int)::Float64

Description

This function extrapolates the results of the Simulation Extrapolation (SIMEX) method to a new value of the parameter λ = -1.

Arguments

λ::Vector: A vector of the values of the parameter λ.

betas::Matrix: A matrix of the estimated parameters for each value of λ.

errorvarindex::Int: The index of the error variable in the X matrix.

Returns

Float64: The extrapolated value of the parameter.

References

Cook, John R., and Leonard A. Stefanski. "Simulation-extrapolation estimation in parametric

measurement error models." Journal of the American Statistical association 89.428 (1994): 1314-1328.